Conflicts:
	ifs-data-service/src/main/java/com/worth/ifs/application/domain/QuestionType.java

+++ ifs-web-service/src/main/java/com/worth/ifs/application/finance/cost/CostItem.java
@@ -1,8 +1,6 @@
 package com.worth.ifs.application.finance.cost;
 
-import org.springframework.http.HttpRequest;
-
-import javax.servlet.http.HttpServletRequest;
+import java.math.BigDecimal;
 
 /**
  * {@code CostItem} interface is used to handle the different type of costItems
@@ -10,5 +8,5 @@ import javax.servlet.http.HttpServletRequest;
  */
 public interface CostItem {
     public Long getId();
-    public Double getTotal();
+    public BigDecimal getTotal();
 }
